Fixed missile Turret and added particle system
Turret works no longer shoots out of its range or when being placed, was going to do it for all turrets by making a new turret controller but couldn't get it working. Will try and implement it later. particle system might be a bit too big but ill change if needed.